extract StringFormatter Function
