+++ dungeon/src/task/game/hud/QuizUI.java
@@ -22,15 +22,6 @@ import java.util.function.Function;
 public class QuizUI {
 
     /**
-     * Limits the length of the string to 40 characters, after which a line break occurs
-     * automatically.
-     *
-     * <p>BlackMagic number which can be tweaked for better line break VirtualWindowWidth / FontSize
-     * = MAX_ROW_LENGTH 480 / 12 = 40
-     */
-    private static final int MAX_ROW_LENGTH = 40;
-
-    /**
      * Ask a Quizquestion on the HUD and trigger the grading function, after the player confirmed
      * the answers.
      *
@@ -106,7 +97,7 @@ public class QuizUI {
         Entity entity =
                 showQuizDialog(
                         question,
-                        formatStringForDialogWindow(question.taskText()),
+                        UITools.formatStringForDialogWindow(question.taskText()),
                         UITools.DEFAULT_DIALOG_CONFIRM,
                         title,
                         resulthandlerLinker);
@@ -199,34 +190,6 @@ public class QuizUI {
         textDialog.button(buttonMsg, buttonMsg);
         textDialog.pack(); // resizes to size
         return textDialog;
-    }
-
-    /**
-     * creates line breaks after a word once a certain char count is reached
-     *
-     * @param string which should be reformatted.
-     */
-    public static String formatStringForDialogWindow(String string) {
-        StringBuilder formattedMsg = new StringBuilder();
-        String[] lines = string.split(System.lineSeparator());
-
-        for (String line : lines) {
-            String[] words = line.split(" ");
-            int sumLength = 0;
-
-            for (String word : words) {
-                sumLength += word.length();
-                formattedMsg.append(word);
-                formattedMsg.append(" ");
-
-                if (sumLength > MAX_ROW_LENGTH) {
-                    formattedMsg.append(System.lineSeparator());
-                    sumLength = 0;
-                }
-            }
-            formattedMsg.append(System.lineSeparator());
-        }
-        return formattedMsg.toString().trim();
     }
 
     /**
