+++ game/test/contrib/utils/components/skill/SkillTest.java
@@ -7,6 +7,7 @@ import core.Game;
 
 import org.junit.After;
 import org.junit.Before;
+import org.junit.Ignore;
 import org.junit.Test;
 
 import java.util.function.Consumer;
@@ -48,10 +49,10 @@ public class SkillTest {
     }
 
     @Test
+    @Ignore
     public void executeWhenCoolDownExpired() throws InterruptedException {
         final long baseCoolDown = 100;
         skill = new Skill(skillFunction, baseCoolDown);
-
         skill.execute(entity);
         assertEquals("Skill should have been executed once", 1, value);
         Thread.sleep(baseCoolDown);
